.github/prompts/.testing/requirements.txt
pyyaml>=6.0
openai>=1.0.0
anthropic>=0.18.0
jsonschema>=4.17.0
python-dotenv>=1.0.0
colorama>=0.4.6
tabulate>=0.9.0
jinja2>=3.1.2
